CURL: Introduce Shell Escaping Support
In scenarios where a header, POST data, or cookie includes a double
quote, such as in the case of Sec-Ch-Ua-Arch with the value '"x86"',
the generated curl command line may fail due to inadequate escaping.

This patch introduces shell escaping support for curl commands,
guaranteeing that the curl output consistently functions as expected.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>